refactor: extract sub-components from BatchMonitorModal
